fix: make i18n in config file possible.
Due to limitation of docusarus, it is quite hard to use i18n tools
in bootstrapping, leading to difficulty of making i18n in config.

So, I replace the used method that mount the config in docusarus
custom fields with directly using the config file.

It has been tested, and it works well.